books_app/requirements.txt
pandas>=1.3.0
google-generativeai>=0.3.0
scikit-learn>=1.0.0
numpy>=1.21.0